Refactored files organization
Refactored the whole application in order to standardize folder organization #9.
