+++ card-game-client/src/components/ActiveGame/ActiveGame.css
@@ -1,0 +1,30 @@
+.game-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.controls-panel {
+  flex: 1;
+  min-width: 300px;
+  background-color: #3a3f4a;
+  padding: 20px;
+  border-radius: 8px;
+  align-self: flex-start;
+}
+
+.controls-panel h2,
+.controls-panel h3 {
+  margin-top: 0;
+  border-bottom: 1px solid #555;
+  padding-bottom: 10px;
+}
+
+.state-panel {
+  flex: 2;
+  min-width: 400px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
