+++ card-game-client/src/components/shared/CardComponent.css
@@ -1,0 +1,8 @@
+.card {
+  background-color: #f0f0f0;
+  color: #282c34;
+  padding: 5px 10px;
+  border-radius: 4px;
+  font-size: 0.9em;
+  font-weight: bold;
+}
