+++ card-game-client/src/components/shared/CardComponent.tsx
@@ -1,0 +1,14 @@
+import { Card } from "../../types/api";
+import './CardComponent.css'
+
+type CardProps = {
+  card: Card;
+};
+
+export function CardComponent({card}: CardProps): React.ReactElement {
+  return (
+    <span className="card">
+      {card.rank} of {card.suit}
+    </span>
+  );
+}
